+++ paddle_billing/Entities/CustomerPortalSessions/CustomerPortalSessionGeneralUrl.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+from dataclasses import dataclass
+
+
+@dataclass
+class CustomerPortalSessionGeneralUrl:
+    overview: str
+
+    @staticmethod
+    def from_dict(data: dict) -> CustomerPortalSessionGeneralUrl:
+        return CustomerPortalSessionGeneralUrl(
+            overview=data["overview"],
+        )